LPS-77447 Move com.liferay.mobile.device.rules.internal.upgrade to web module
